+++ packages/app/src/api/core-domain/infra/db/raw.ts
@@ -1,3 +1,4 @@
+import type { ErrorDetailTuple } from "@common/core-domain/domain/ErrorDetailTuple";
 import type { DeclarationDTO as DeclarationDataRaw } from "@common/models/generated";
 
 export interface DeclarationRaw {
@@ -24,11 +25,11 @@ export interface RepresentationEquilibreeRaw {
 export interface OwnershipRequestRaw {
   asker_email: string;
   created_at: string;
-  email: string;
-  error_detail?: string;
+  email: string | null;
+  error_detail: ErrorDetailTuple | null;
   id: string;
   modified_at: string;
-  siren: string;
+  siren: string | null;
   status: string;
 }